requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
Django==4.1.4
django-multiselectfield==0.1.12
easy-thumbnails==2.8.4
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
fontawesomefree==6.6.0